refactor(frontend): phase C — wiring + cleanup
Axios:
- axiosInstance: read VITE_API_URL with fallback to localhost:3000
- Victim helper: drop duplicate axios.create, import shared Axios

Dead deps:
- remove react-hook-form, @hookform/resolvers, zod (zero usage)
- refresh pnpm-lock.yaml

Error boundary:
- new components/ErrorBoundary.jsx (render-phase only)
- main.jsx wraps RouterProvider

File rename hygiene (git mv to preserve blame):
- utils/checkAuth.jsx -> utils/ProtectedRoute.jsx
- api/authService.jsx -> api/authService.js (no JSX)
- api/Victim.jsx -> api/victim.js (no JSX)
- update all import callers (main, UserDetail, BoxingRing)

Auth.jsx: add note comment pointing at authService.js as canonical layer
(consolidation deferred per Standard scope).

Phase B follow-up:
- remove dead getAllVictim from victim.models
- remove dead getAllHitEffects from hitEffect.model
