+++ frontend/src/main.jsx
@@ -8,9 +8,10 @@ import Auth from './pages/Auth.jsx';
 import UserDetail from './pages/UserDetail.jsx';
 import BoxingRing from './pages/BoxingRing.jsx';
 import NotFoundPage from './pages/NotFoundPage.jsx';
-import ProtectedRoute from "./utils/checkAuth.jsx";
+import ProtectedRoute from "./utils/ProtectedRoute.jsx";
 import FuneralPopup from "./components/FuneralPopup.jsx";
 import Forgive from "./components/Forgive.jsx";
+import ErrorBoundary from "./components/ErrorBoundary.jsx";
 
 const router = createBrowserRouter([
     {
@@ -53,6 +54,8 @@ const router = createBrowserRouter([
 
 createRoot(document.getElementById('root')).render(
     <StrictMode>
-        <RouterProvider router={router}/>
+        <ErrorBoundary>
+            <RouterProvider router={router}/>
+        </ErrorBoundary>
     </StrictMode>,
 )
