+++ frontend/src/utils/axiosInstance.js
@@ -1,7 +1,7 @@
 import axios from 'axios';
 
 const Axios = axios.create({
-	baseURL: 'http://localhost:3000',
+	baseURL: import.meta.env.VITE_API_URL || 'http://localhost:3000',
 	withCredentials: true, // This is important for cookies
 	headers: {
 		'Content-Type': 'application/json',
